+++ CHANGELOG.md
@@ -443,7 +443,7 @@ this release.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/thriftrw/thriftrw-go/compare/v1.31.0...HEAD
+[1.32.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.31.0...v1.32.0
 [1.31.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.30.0...v1.31.0
 [1.30.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.29.2...v1.30.0
 [1.29.2]: https://github.com/thriftrw/thriftrw-go/compare/v1.29.1...v1.29.2